Fix broken Sparkfun/Huzzah "Get Started" links
It appears the Sparkfun and Huzzah "Get Started" links are 404'd, perhaps because these samples are obsolete. To start the conversation let's link to the archived Github repositories instead, which should explain why they're obsolete ideally. (Perhaps it'd be more appropriate to just remove points 8 and 9 here, remove the links entirely.)